% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answ_freq_table.R
\name{answ_freq_table}
\alias{answ_freq_table}
\title{Create a frequency table of answers options}
\usage{
answ_freq_table(
  data,
  vars,
  expand_to,
  relevel_by,
  recode_to_german,
  recode_to_french,
  recode_to_italian,
  set_colors
)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{a vector passing a selection of variables}

\item{expand_to}{specify all factor levels (numeric), arguments are passed to fct_expand}

\item{relevel_by}{specify order of factor levels, arguments passed to fct_relevel}

\item{recode_to_german}{specify (alphanumeric) answer options, arguments passed to fct_recode,
example: c("Neuer Text" = "1", "Zweiter Text" = "2")}

\item{recode_to_french}{specify (alphanumeric) answer options, arguments passed to fct_recode}

\item{recode_to_italian}{specify (alphanumeric) answer options, arguments passed to fct_recode}

\item{set_colors}{specify coloring by (numeric) answer, arguments passed to fct_recode}
}
\value{
a dataframe
}
\description{
Create a frequency table of answers options
}
\section{Helpers}{
NA
}

