% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_redcap_username.R
\name{complete_redcap_username}
\alias{complete_redcap_username}
\title{Complete redcap usernames based on csv-filled-lists with names and email addresses}
\usage{
complete_redcap_username(which_file = "latest")
}
\arguments{
\item{which_file}{as standard REDCap user name list filled out with names and email addresses}
}
\value{
a csv-file with suffix _completed.csv
}
\description{
Complete redcap usernames based on csv-filled-lists with names and email addresses
}
\details{
Admins have the rights to add additional users to REDCap. The PI or data managers of a
project is responsible to fill out the user form and send it via Email to the admin.
The admin fills out the usernames according to a standardized rule (handle of the email-adress).
When bulk-adding multiple users, this is tedious. This function fills completes the
user names automatically and saves the file as a .csv in the same directory as the
supplied file.
}
