% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-funs.R
\name{get_model_dotplot}
\alias{get_model_dotplot}
\title{Create ranef-dotplot from model}
\usage{
get_model_dotplot(model)
}
\arguments{
\item{model}{a \code{glmer()}-object from class S4}
}
\value{
a plot
}
\description{
Create ranef-dotplot from model
}
