% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-funs.R
\name{get_model_table}
\alias{get_model_table}
\title{create a tidy table out of the glmer-object to plot}
\usage{
get_model_table(model)
}
\arguments{
\item{model}{a \code{glmer()}-object from class S4}
}
\value{
a tibble with random effects and random coefficients
}
\description{
create a tidy table out of the glmer-object to plot
}
