% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-palette.R
\name{get_unibas_palettes}
\alias{get_unibas_palettes}
\title{Get names of all unique palettes provided in unibas}
\usage{
get_unibas_palettes(full = FALSE)
}
\arguments{
\item{full}{Whether to include full palette names (with suffixes, e.g. \code{_cont}) or just stubs}
}
\value{
Vector of palette name stubs or full names
}
\description{
Get names of all unique palettes provided in unibas
}
\keyword{internal}
