% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-funs.R
\name{ggCaterpillar}
\alias{ggCaterpillar}
\title{ggCaterpillar plot from random effects model}
\source{
from \url{https://stackoverflow.com/questions/13847936/plot-random-effects-from-lmer-lme4-package-using-qqmath-or-dotplot-how-to-mak}
}
\usage{
ggCaterpillar(model, condVar = TRUE, QQ = FALSE, likeDotplot = TRUE)
}
\arguments{
\item{model}{glmer-model to extract the random effects out}

\item{condVar}{input to \code{lme4::ranef(model, condVar)}, default = T}

\item{QQ}{TRUE/FALSE, default = FALSE}

\item{likeDotplot}{TRUE/FALSE, default = FALSE}
}
\description{
ggCaterpillar plot from random effects model
}
