% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggStackfreq.R
\name{ggStackfreq}
\alias{ggStackfreq}
\title{Create a customized stacked barplot with percentage-frequencies}
\usage{
ggStackfreq(
  data,
  vars,
  expand_to,
  relevel_by,
  recode_to_german,
  recode_to_french,
  recode_to_italian,
  set_colors,
  language
)
}
\arguments{
\item{data}{a data frame as result of answ_freq_table}

\item{vars}{a vector passing a selection of variables}

\item{expand_to}{specify all factor levels (numeric), arguments are passed to fct_expand}

\item{relevel_by}{specify order of factor levels, arguments passed to fct_relevel}

\item{recode_to_german}{specify (alphanumeric) answer options, arguments passed to fct_recode, example: c("Neuer Text" = "1", "Zweiter Text" = "2")}

\item{recode_to_french}{specify (alphanumeric) answer options, arguments passed to fct_recode}

\item{recode_to_italian}{specify (alphanumeric) answer options, arguments passed to fct_recode}

\item{set_colors}{specify coloring by (numeric) answer, arguments passed to fct_recode}

\item{language}{one of c("german, "french", "italian"), default == "german", abbreviations tolerated}
}
\value{
a ggplot-object
}
\description{
Create a customized stacked barplot with percentage-frequencies
}
\keyword{internal}
