% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_agree.R
\name{n_agree}
\alias{n_agree}
\title{get number of agreement answers (defined by a cutoff) in a vector}
\usage{
n_agree(x, cutoff = 1)
}
\arguments{
\item{x}{a numeric/integer vector}

\item{cutoff}{a vector defining positive answers, default = 1 in a binary dataset}
}
\value{
a numeric value
}
\description{
get number of agreement answers (defined by a cutoff) in a vector
}
\examples{
# usage with dichotomized vector
vec <- c(0, 0, 1, 1, 1)
n_agree(vec)

# usage with non-dichotomized vector:
vec <- c(1, 2, 3, 4, 3, 3, 4, 3, 4, 4, NA)
# you need to change the cutoff
n_agree(vec, cutoff = c(3,4))



}
