% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_missing.R
\name{n_missing}
\alias{n_missing}
\title{get number of NAs in a vector (complement to n_valid())}
\usage{
n_missing(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a numeric vector
}
\description{
get number of NAs in a vector (complement to n_valid())
}
