% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_valid.R
\name{n_valid}
\alias{n_valid}
\title{get sum of valid (not NAs) of a vector}
\usage{
n_valid(x)
}
\arguments{
\item{x}{a vector}
}
\value{
an integer value
}
\description{
get sum of valid (not NAs) of a vector
}
