% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_agree.R
\name{prop_agree}
\alias{prop_agree}
\title{get (proportion) percentage agreement of answers in a vector - 'agreement' to be defined
by cutoff}
\usage{
prop_agree(x, cutoff = 1, restrict_level = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{cutoff}{a vector defining positive answers, default = 1 in a binary dataset}

\item{restrict_level}{the minimum number of valid answers needed to display result,
otherwise the answer is restricted and set to NA, default is NULL (no restriction)}
}
\value{
a numeric value between 0 - 1
}
\description{
get (proportion) percentage agreement of answers in a vector - 'agreement' to be defined
by cutoff
}
\examples{
# binary vector
vec <- c(0, 1, 1, 0, 0, 1, NA_real_, 1)
prop_agree(x = vec, cutoff = 1)

# with restriction - throws NA as not applicable
prop_agree(x = vec, cutoff = 1, restrict_level = 8)
}
