% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-palette.R
\name{scale_custom}
\alias{scale_custom}
\alias{scale_color_unibas_d}
\alias{scale_colour_unibas_d}
\alias{scale_fill_unibas_d}
\alias{scale_color_unibas_op}
\alias{scale_colour_unibas_op}
\alias{scale_fill_unibas_op}
\alias{scale_color_unibas_div}
\alias{scale_colour_unibas_div}
\alias{scale_fill_unibas_div}
\alias{scale_color_unibas_c}
\alias{scale_colour_unibas_c}
\alias{scale_fill_unibas_c}
\title{Custom color and fill scales}
\usage{
scale_color_unibas_d(palette = "unibas", extend = FALSE, ...)

scale_colour_unibas_d(palette = "unibas", extend = FALSE, ...)

scale_fill_unibas_d(palette = "unibas", extend = FALSE, ...)

scale_color_unibas_op(
  palette = "unibas",
  val_names = c("good", "neutral", "bad"),
  ...
)

scale_colour_unibas_op(
  palette = "unibas",
  val_names = c("good", "neutral", "bad"),
  ...
)

scale_fill_unibas_op(
  palette = "unibas",
  val_names = c("good", "neutral", "bad"),
  ...
)

scale_color_unibas_div(palette = "unibas", ...)

scale_colour_unibas_div(palette = "unibas", ...)

scale_fill_unibas_div(palette = "unibas", ...)

scale_color_unibas_c(palette = "unibas", ...)

scale_colour_unibas_c(palette = "unibas", ...)

scale_fill_unibas_c(palette = "unibas", ...)
}
\arguments{
\item{palette}{Name of color palette}

\item{extend}{Whether to extend discrete color palette to make sufficient colors for levels needed}

\item{...}{Additional arguments to be passed to internal scale function}

\item{val_names}{For opinionated scales, defaults to "good", "neutral", "bad"}
}
\description{
Custom coloring and filling functions based on unique color palettes
}
\details{
Specific functions include:
\itemize{
\item{\code{scale_(color/colour/fill)_unibas_d}
{: Discrete palette with either fixed or dynamically extended number of shades}}
\item{\code{scale_(color/colour/fill)_unibas_op}\\b
{: Discrete palette with fixed colors for "good", "bad", and "neutral"}}
\item{\code{scale_(color/colour/fill)_unibas_div}
{: Continuous diverging color palette, must contain negative, neutral, positive values}}
\item{\code{scale_(color/colour/fill)_unibas_c}
{: Continuous color palette}}
}
}
\examples{
library(ggplot2)
library(hsrtools)
ggplot(diamonds[1:2000,], aes(x = cut, y = carat,
color = cut)) +
geom_point() +
scale_color_unibas_d()
}
