% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibas_theme.R
\name{theme_unibas}
\alias{theme_unibas}
\title{unibas ggplot theme}
\usage{
theme_unibas(base_theme = ggplot2::theme_minimal(), base_size = 12, ...)
}
\arguments{
\item{base_theme}{A base theme upon which additional theme-specific options are applied}

\item{base_size}{Base font size, defaults to 12, passed to theme}

\item{...}{Further arguments passed to \code{ggplot2::theme()}}
}
\description{
\link{ggplot2} plot theme based on colors of UNIBAS
}
\examples{
library(ggplot2)
df <- data.frame(x = as.factor(1:8),
                 y = c(300, 500, 600, 750, 700, 550, 450, 350))

ggplot(df, aes(x, y, fill = x)) +
  geom_col() +
  scale_fill_unibas_d() +
  theme_unibas()

}
\seealso{
\href{https://ggplot2.tidyverse.org/reference/theme.html}{\code{ggplot2::theme()}}
}
