% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibas_color_palette.R
\docType{data}
\name{unibas_color_palette_overview}
\alias{unibas_color_palette_overview}
\title{Color palette proposed by Unibas CI}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
unibas_color_palette_overview
}
\description{
Overview all colors from Nursing CI info sheet (found on ADAM)
}
\examples{
scales::show_col(unibas_color_palette_overview$hex_value)
}
\keyword{datasets}
