% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibas_pals.R
\docType{data}
\name{unibas_pal}
\alias{unibas_pal}
\alias{unibas_pal_op}
\alias{unibas_pal_cont}
\alias{unibas_pal_div}
\title{Unibas colour palette to use in figures}
\format{
An object of class \code{character} of length 10.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.
}
\usage{
unibas_pal

unibas_pal_op

unibas_pal_cont

unibas_pal_div
}
\description{
Unibas colour palette to use in figures
}
\keyword{datasets}
