% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibas_color_palette.R
\docType{data}
\name{unibas_palette}
\alias{unibas_palette}
\alias{unibas_pal_op}
\alias{unibas_pal_cont}
\alias{unibas_pal_div}
\title{Unibas colour palette to use in figures}
\format{
An object of class \code{character} of length 10.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.
}
\usage{
unibas_palette

unibas_pal_op

unibas_pal_cont

unibas_pal_div
}
\description{
A vector holding the information from \code{unibas_color_palette_overview$hex_value}
}
\examples{
library(ggplot2)
ggplot(diamonds, aes(x = cut, y = carat)) +
geom_col(aes(fill = cut)) +
scale_fill_manual(values = unibas_palette) +
 labs(x = "",
  title = "Diamonds: Carat by Cut",
  subtitle = "in the Unibas color-palette") +
  theme_minimal()
}
\keyword{datasets}
