% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_created_on.R
\name{file_created_on}
\alias{file_created_on}
\title{Adds 'file_created_on'-variable to a tidy dataframe}
\usage{
file_created_on(.data, var_name = "file_created_on")
}
\arguments{
\item{.data}{a tidy dataframe}

\item{var_name}{name of the variable to be added, defaults to "file_created_on"}
}
\value{
a tidy dataframe
}
\description{
Adds 'file_created_on'-variable to a tidy dataframe
}
\examples{

library(magrittr)
# sample dataframe
df1 <- tibble::tibble(ID = seq_len(100), var1 = sample(c(1:3), 100, TRUE))

# add variable
file_created_on(df1)
}
