% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_format.R
\name{journal_format}
\alias{journal_format}
\title{Format numbers according to journal requirements}
\usage{
journal_format(x, n_break = 4, ...)
}
\arguments{
\item{x}{an integer}

\item{n_break}{fix number where the formatting sets the break}

\item{...}{additional arguments passed to base::format()}
}
\value{
an formatted character string
}
\description{
Format numbers according to journal requirements
}
\keyword{internal}
