% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unibas_color_palette_overview}
\alias{unibas_color_palette_overview}
\title{unibas_color_palette_overview}
\format{
\subsection{\code{unibas_color_palette_overview}}{

A data frame with 10 colors, in name, rgb_value and hex.
}
}
\usage{
unibas_color_palette_overview
}
\description{
unibas_color_palette_overview
}
\keyword{datasets}
