% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibas_color_palette.R
\docType{data}
\name{unibas_color_palette_overview}
\alias{unibas_color_palette_overview}
\title{Overview of color palette proposed by unibas corporate design}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
unibas_color_palette_overview
}
\description{
A table giving an overview all colors (taken from the CI info sheet on ADAM)
with both RGB and Hex values
}
\examples{
scales::show_col(unibas_color_palette_overview$hex_value)
}
\keyword{datasets}
