% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibas_color_palette.R
\docType{data}
\name{unibas_palette}
\alias{unibas_palette}
\title{Colour palette to use in figures}
\format{
An object of class \code{character} of length 10.
}
\usage{
unibas_palette
}
\description{
A vector holding the information from \code{unibas_color_palette_overview$hex_value}
}
\keyword{datasets}
